Some changes

+++ src/engine/engine.mm
@@ -31,16 +31,12 @@ static struct EngineState {
     bool minimized = false;
 } state;
 
-static void default_callback() {}
-static void default_window_resize_callback(uint32_t, uint32_t, uint32_t, uint32_t) {}
-static bool default_load_assets_callback() { return true; }
-
-static void handle_keyboard_events(GLFWwindow* window, int key, int scancode, int action, int mods);
-static void handle_mouse_button_events(GLFWwindow* window, int button, int action, int mods);
-static void handle_mouse_scroll_events(GLFWwindow* window, double xoffset, double yoffset);
-static void handle_cursor_pos_events(GLFWwindow* window, double xpos, double ypos);
-static void handle_window_resize_events(GLFWwindow* window, int width, int height);
-static void handle_window_iconify_callback(GLFWwindow* window, int iconified);
+static void HandleKeyboardEvents(GLFWwindow* window, int key, int scancode, int action, int mods);
+static void HandleMouseButtonEvents(GLFWwindow* window, int button, int action, int mods);
+static void HandleMouseScrollEvents(GLFWwindow* window, double xoffset, double yoffset);
+static void HandleCursorPosEvents(GLFWwindow* window, double xpos, double ypos);
+static void HandleWindowResizeEvents(GLFWwindow* window, int width, int height);
+static void HandleWindowIconifyCallback(GLFWwindow* window, int iconified);
 
 static inline const char* glfwGetErrorString() {
     const char* description = nullptr;
@@ -48,7 +44,7 @@ static inline const char* glfwGetErrorString() {
     return description;
 }
 
-static GLFWwindow* create_window(LLGL::Extent2D size, bool fullscreen, bool hidden, uint8_t samples) {
+static GLFWwindow* CreateWindow(LLGL::Extent2D size, const char* title, bool fullscreen, bool hidden, uint8_t samples) {
     glfwWindowHint(GLFW_FOCUSED, 1);
     glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
     glfwWindowHint(GLFW_VISIBLE, hidden ? GLFW_FALSE : GLFW_TRUE);
@@ -56,18 +52,18 @@ static GLFWwindow* create_window(LLGL::Extent2D size, bool fullscreen, bool hidd
 
     GLFWmonitor* primary_monitor = fullscreen ? glfwGetPrimaryMonitor() : nullptr;
 
-    GLFWwindow *window = glfwCreateWindow(size.width, size.height, "AAA", primary_monitor, nullptr);
+    GLFWwindow *window = glfwCreateWindow(size.width, size.height, title, primary_monitor, nullptr);
     if (window == nullptr) {
         SGE_LOG_ERROR("Couldn't create a window: %s", glfwGetErrorString());
         return nullptr;
     }
 
-    glfwSetKeyCallback(window, handle_keyboard_events);
-    glfwSetMouseButtonCallback(window, handle_mouse_button_events);
-    glfwSetScrollCallback(window, handle_mouse_scroll_events);
-    glfwSetCursorPosCallback(window, handle_cursor_pos_events);
-    glfwSetWindowSizeCallback(window, handle_window_resize_events);
-    glfwSetWindowIconifyCallback(window, handle_window_iconify_callback);
+    glfwSetKeyCallback(window, HandleKeyboardEvents);
+    glfwSetMouseButtonCallback(window, HandleMouseButtonEvents);
+    glfwSetScrollCallback(window, HandleMouseScrollEvents);
+    glfwSetCursorPosCallback(window, HandleCursorPosEvents);
+    glfwSetWindowSizeCallback(window, HandleWindowResizeEvents);
+    glfwSetWindowIconifyCallback(window, HandleWindowIconifyCallback);
 
     return window;
 }
@@ -80,52 +76,42 @@ static inline LLGL::Extent2D get_scaled_resolution(uint32_t width, uint32_t heig
 }
 
 void Engine::SetPreUpdateCallback(PreUpdateCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.pre_update_callback = callback;
 }
 
 void Engine::SetUpdateCallback(UpdateCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.update_callback = callback;
 }
 
 void Engine::SetPostUpdateCallback(PostUpdateCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.post_update_callback = callback;
 }
 
 void Engine::SetFixedUpdateCallback(FixedUpdateCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.fixed_update_callback = callback;
 }
 
 void Engine::SetFixedPostUpdateCallback(FixedUpdateCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.fixed_post_update_callback = callback;
 }
 
 void Engine::SetRenderCallback(RenderCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.render_callback = callback;
 }
 
 void Engine::SetPostRenderCallback(PostRenderCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.post_render_callback = callback;
 }
 
 void Engine::SetDestroyCallback(PostRenderCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.destroy_callback = callback;
 }
 
 void Engine::SetWindowResizeCallback(WindowResizeCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.window_resize_callback = callback;
 }
 
 void Engine::SetLoadAssetsCallback(LoadAssetsCallback callback) {
-    SGE_ASSERT(callback != nullptr, "Pointer to the callback must not be null.");
     state.load_assets_callback = callback;
 }
 
@@ -148,17 +134,6 @@ void Engine::SetCursorMode(CursorMode cursor_mode) {
 bool Engine::Init(sge::RenderBackend backend, sge::WindowSettings settings, LLGL::Extent2D& output_viewport) {
     ZoneScopedN("Engine::Init");
 
-    if (state.pre_update_callback == nullptr) state.pre_update_callback = default_callback;
-    if (state.update_callback == nullptr) state.update_callback = default_callback;
-    if (state.post_update_callback == nullptr) state.post_update_callback = default_callback;
-    if (state.fixed_update_callback == nullptr) state.fixed_update_callback = default_callback;
-    if (state.fixed_post_update_callback == nullptr) state.fixed_post_update_callback = default_callback;
-    if (state.render_callback == nullptr) state.render_callback = default_callback;
-    if (state.post_render_callback == nullptr) state.post_render_callback = default_callback;
-    if (state.destroy_callback == nullptr) state.destroy_callback = default_callback;
-    if (state.window_resize_callback == nullptr) state.window_resize_callback = default_window_resize_callback;
-    if (state.load_assets_callback == nullptr) state.load_assets_callback = default_load_assets_callback;
-
 #if SGE_PLATFORM_LINUX
     glfwInitHint(GLFW_PLATFORM, GLFW_PLATFORM_WAYLAND);
 #endif
@@ -178,7 +153,7 @@ bool Engine::Init(sge::RenderBackend backend, sge::WindowSettings settings, LLGL
         window_size = LLGL::Extent2D(mode->width, mode->height);
     }
 
-    GLFWwindow *window = create_window(window_size, settings.fullscreen, settings.hidden, settings.samples);
+    GLFWwindow *window = CreateWindow(window_size, settings.title, settings.fullscreen, settings.hidden, settings.samples);
     if (window == nullptr) return false;
 
     state.window = window;
@@ -188,10 +163,14 @@ bool Engine::Init(sge::RenderBackend backend, sge::WindowSettings settings, LLGL
     output_viewport.width = window_size.width;
     output_viewport.height = window_size.height;
 
+    SetCursorMode(settings.cursor_mode);
+
     const LLGL::Extent2D resolution = get_scaled_resolution(window_size.width, window_size.height);
     if (!state.renderer.Init(window, resolution, settings)) return false;
 
-    if (!state.load_assets_callback()) return false;
+    if (state.load_assets_callback) {
+        if (!state.load_assets_callback()) return false;
+    }
 
     Time::SetFixedTimestepSeconds(1.0f / 60.0f);
 
@@ -212,29 +191,37 @@ void Engine::Run() {
             const delta_time_t dt(delta_time);
             Time::AdvanceBy(dt);
 
-            state.pre_update_callback();
+            if (state.pre_update_callback)
+                state.pre_update_callback();
 
             int fixed_update_count = 0;
 
             fixed_timer += delta_time;
             while (fixed_timer >= Time::FixedDeltaSeconds()) {
                 Time::AdvanceFixed();
-                state.fixed_update_callback();
+                if (state.fixed_update_callback)
+                    state.fixed_update_callback();
                 fixed_timer -= Time::FixedDeltaSeconds();
                 ++fixed_update_count;
             }
 
-            state.update_callback();
+            if (state.update_callback)
+                state.update_callback();
 
             for (int i = 0; i < fixed_update_count; ++i) {
-                state.fixed_post_update_callback();
+                if (state.fixed_post_update_callback)
+                    state.fixed_post_update_callback();
             }
 
-            state.post_update_callback();
+            if (state.post_update_callback)
+                state.post_update_callback();
 
             if (!state.minimized) {
-                state.render_callback();
-                state.post_render_callback();
+                if (state.render_callback)
+                    state.render_callback();
+
+                if (state.post_render_callback)
+                    state.post_render_callback();
             }
 
             Input::Clear();
@@ -249,7 +236,9 @@ void Engine::Destroy() {
         state.renderer.CommandQueue()->WaitIdle();
     }
 
-    state.destroy_callback();
+    if (state.destroy_callback != nullptr) {
+        state.destroy_callback();
+    }
 
     if (state.renderer.Context()) {
         state.renderer.Terminate();
@@ -260,7 +249,7 @@ void Engine::Destroy() {
 
 sge::Renderer& Engine::Renderer() { return state.renderer; }
 
-static void handle_keyboard_events(GLFWwindow*, int key, int, int action, int) {
+static void HandleKeyboardEvents(GLFWwindow*, int key, int, int action, int) {
     if (action == GLFW_PRESS) {
         Input::Press(static_cast<Key>(key));
     } else if (action == GLFW_RELEASE) {
@@ -268,7 +257,7 @@ static void handle_keyboard_events(GLFWwindow*, int key, int, int action, int) {
     }
 }
 
-static void handle_mouse_button_events(GLFWwindow*, int button, int action, int) {
+static void HandleMouseButtonEvents(GLFWwindow*, int button, int action, int) {
     if (action == GLFW_PRESS) {
         Input::Press(static_cast<MouseButton>(button));
     } else if (action == GLFW_RELEASE) {
@@ -276,11 +265,11 @@ static void handle_mouse_button_events(GLFWwindow*, int button, int action, int)
     }
 }
 
-static void handle_mouse_scroll_events(GLFWwindow*, double, double yoffset) {
+static void HandleMouseScrollEvents(GLFWwindow*, double, double yoffset) {
     Input::PushMouseScrollEvent(yoffset);
 }
 
-static void handle_cursor_pos_events(GLFWwindow*, double xpos, double ypos) {
+static void HandleCursorPosEvents(GLFWwindow*, double xpos, double ypos) {
     const glm::uvec2 window_size = glm::uvec2(state.window_width, state.window_height);
 
     xpos = std::min(std::max(xpos, 0.0), static_cast<double>(window_size.x));
@@ -289,7 +278,7 @@ static void handle_cursor_pos_events(GLFWwindow*, double xpos, double ypos) {
     Input::SetMouseScreenPosition(glm::vec2(xpos, ypos));
 }
 
-static void handle_window_resize_events(GLFWwindow*, int width, int height) {
+static void HandleWindowResizeEvents(GLFWwindow*, int width, int height) {
     if (width <= 0 || height <= 0) {
         state.minimized = true;
         return;
@@ -304,11 +293,13 @@ static void handle_window_resize_events(GLFWwindow*, int width, int height) {
     state.window_width = width;
     state.window_height = height;
 
-    state.window_resize_callback(static_cast<uint32_t>(width), static_cast<uint32_t>(height), resolution.width, resolution.height);
+    if (state.window_resize_callback != nullptr) {
+        state.window_resize_callback(static_cast<uint32_t>(width), static_cast<uint32_t>(height), resolution.width, resolution.height);
+    }
 
     state.render_callback();
 }
 
-static void handle_window_iconify_callback(GLFWwindow*, int iconified) {
+static void HandleWindowIconifyCallback(GLFWwindow*, int iconified) {
     state.minimized = iconified;
 }
